Refactor autotrack prop logic into 'common' file
